requirements.txt
asgiref==3.4.1
Django==3.2.16
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
Pillow==8.4.0
PyJWT==1.7.1
pytz==2022.5
simplejwt==2.0.1
sqlparse==0.4.3
typing==3.7.4.3
typing-extensions==4.1.1
